refactor: vo 패키지 수정
